pyproject.toml
[tool.poetry]
name = "yelprecommendation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "2.2.2"
torchaudio = "2.2.2"
torchvision = "0.17.2"
pandas = "^2.2.2"
tqdm = "^4.66.2"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
hydra-core = "^1.3.2"
loguru = "^0.7.2"
scikit-learn = "1.4.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
